% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_pal}
\alias{get_pal}
\title{Select shoalsmaRine lab palette}
\usage{
get_pal(palette_name)
}
\arguments{
\item{palette_name}{Name of desired palette}
}
\value{
A vector of hex colour codes
}
\description{
This function returns colour palettes based on shoals marine lab photos
}
\examples{
get_pal("sunset")
get_pal("gilly")
}
