% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{print_pal}
\alias{print_pal}
\title{Plot colour palette}
\usage{
print_pal(x)
}
\arguments{
\item{x}{A vector of colours}
}
\value{
An image showing colours in palette
}
\description{
Plot chosen colour palette from shoalsmaRine lab package
}
\examples{
pardalote <- get_pal("sunset")
print_pal(sunset)
}
