% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_sml}
\alias{scale_color_sml}
\title{ggplot2 color scale using Shoals Marine Lab palettes}
\usage{
scale_color_sml(palette = "gull", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of the palette.}

\item{discrete}{Logical. Should the scale be discrete?}

\item{reverse}{Logical. Should the palette order be reversed?}

\item{...}{Additional arguments passed to ggplot2 scale functions.}
}
\value{
A ggplot2 scale.
}
\description{
ggplot2 color scale using Shoals Marine Lab palettes
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point(size = 3) +
  scale_color_sml("sunset")
}
