% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_sml}
\alias{scale_fill_sml}
\title{ggplot2 fill scale using Shoals Marine Lab palettes}
\usage{
scale_fill_sml(palette = "gull", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of the palette.}

\item{discrete}{Logical. Should the scale be discrete?}

\item{reverse}{Logical. Should the palette order be reversed?}

\item{...}{Additional arguments passed to ggplot2 scale functions.}
}
\value{
A ggplot2 scale.
}
\description{
ggplot2 fill scale using Shoals Marine Lab palettes
}
\examples{
library(ggplot2)

ggplot(iris, aes(Species, Sepal.Length, fill = Species)) +
  geom_boxplot() +
  scale_fill_sml("sunset")
}
