% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{sml_palette}
\alias{sml_palette}
\title{Return a Shoals Marine Lab color palette}
\usage{
sml_palette(name, n = NULL, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of the palette.}

\item{n}{Number of colors to return. If \code{NULL}, returns all colors.}

\item{type}{Either \code{"discrete"} or \code{"continuous"}.}
}
\value{
A character vector of hex colors.
}
\description{
Return a Shoals Marine Lab color palette
}
\examples{
sml_palette("sunset")
sml_palette("trawl", n = 4)
sml_palette("sunset", n = 5, type = "continuous")
}
