% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{sml_palettes_available}
\alias{sml_palettes_available}
\title{List available Shoals Marine Lab palettes}
\usage{
sml_palettes_available()
}
\value{
A character vector of available palette names.
}
\description{
List available Shoals Marine Lab palettes
}
\examples{
sml_palettes_available()
}
